requirements.txt
# application
flask
flask_json
invoke
pyyaml

# test
pytest
pytest-flask
